Feature/team vms resources

+++ resource_server/src/main/java/it/polito/ai/lab2/services/CourseServiceImpl.java
@@ -232,13 +232,15 @@ public class CourseServiceImpl implements CourseService {
       }
     }
 
-    Path oldVmModelPath = Utility.VM_MODELS_DIR.resolve(c.getVmModel().getId().toString());
-    try {
-      Files.delete(oldVmModelPath);
-    } catch (IOException e) {
-      throw new RuntimeException("Cannot delete the file: " + e.getMessage());
+    if(c.getVmModel() != null) {
+      Path oldVmModelPath = Utility.VM_MODELS_DIR.resolve(c.getVmModel().getId().toString());
+      try {
+        Files.delete(oldVmModelPath);
+      } catch (IOException e) {
+        throw new RuntimeException("Cannot delete the file: " + e.getMessage());
+      }
+      vmModelRepository.delete(c.getVmModel());
     }
-    vmModelRepository.delete(c.getVmModel());
 
     VmModel v = new VmModel();
     v.setName(c.getAcronym());
